runj-entrypoint: new helper program for STDIO
With this commit, runj no longer hooks up the container process STDIO to
the STDIO of `runj start`.  Instead, runj now copies runc's behavior of
hooking up stdio to the `create` subcommand.

This requires a new helper process that is started at the time of
container creation and is responsible for invoking the container
process when `runj start` is eventually called.

Synchronization between `runj-entrypoint` and `runj start` is through a
FIFO (see mkfifo(2)) in runj's state directory.