shim: implement Wait API
The Wait API is used to track process exits from inside the jail.
Enabling Wait requires registering the shim as a reaper for orphaned
processes inside the jail and recording their exit information.
containerd's reaper package and Monitor implementation are used to
provide a subscription model for process exit information.