shim: filter incompatible Linux mounts
In current versions of containerd, the default generated OCI bundle
includes Linux-specific mounts that are incompatible with FreeBSD.  Now
that runj implements mounts, these need to be filtered out.  An upstream
change to containerd addresses the OCI bundle generation logic, but this
change is not yet available in a release.