version: load VERSION and REV_OVERRIDE from file
Go's buildinfo does not include the main module tags, so instead use a
file.  This requires careful management of the VERSION file such that
only a single commit contains the version corresponding to a tag (or
makes it annoying to match up).  The git revision is included to help
correlate commit to tag in case a mistake is made.  The REV_OVERRIDE
file can subsitute commit information when runj is built outside a git
checkout.